requirements.txt
Django==2.2
djangorestframework==3.12.2
django-environ==0.4.5
django-extensions==3.1.5
requests==2.25.1
django-cors-headers==3.7.0
coreapi==2.3.3
pytest-django==4.5.0